+++ packages/terra-framework-docs/CHANGELOG.md
@@ -6,6 +6,9 @@
 * Added
   * Added example for `terra-menu` with disabled item.
 
+* Added
+  * Added a `terra-slide-panel` test containing multiple slide panels.
+
 ## 1.21.0 - (May 11, 2023)
 
 * Changed